+++ pages/changelog.md
@@ -10,6 +10,7 @@ sidebar: auto
 - 修复 `cqhttp` secret 校验出错
 - 修复 `pydantic 1.8` 导致的 `alias` 问题
 - 修改 `cqhttp` `ding` `session id`，不再允许跨群
+- 修改 `shell_command` 存储 message
 
 ## v2.0.0a11